completely rewritten material system
