+++ material/basic.go
@@ -5,15 +5,15 @@ import "github.com/go-gl/gl/v4.1-core/gl"
 type BasicMaterial struct {
 	shader *ShaderProgram
 
-	diffuseLevel float32
+	DiffuseLevel float32
 
-	color [3]float32
+	Hue [4]float32
 
-	diffuseMap      *uint32
-	diffuseMapScale float32
+	DiffuseMap      *uint32
+	DiffuseMapScale float32
 
-	alphaLevel float32
-	alphaMap   *uint32
+	AlphaMapLevel float32
+	AlphaMap      *uint32
 
 	animationPlaying  string
 	animationTextures map[string][]*uint32
@@ -23,15 +23,16 @@ type BasicMaterial struct {
 	animationEnabled  bool
 }
 
-func NewBasicMaterial(shader *ShaderProgram) BasicMaterial {
-	return BasicMaterial{
+func NewBasicMaterial(shader *ShaderProgram) *BasicMaterial {
+	return &BasicMaterial{
 		shader:            shader,
-		diffuseLevel:      0,
-		color:             [3]float32{46, 46, 46},
-		diffuseMapScale:   1,
-		alphaLevel:        1,
+		DiffuseLevel:      0,
+		Hue:               [4]float32{200, 200, 200, 255},
+		DiffuseMapScale:   1,
+		AlphaMapLevel:     0,
 		animationTextures: make(map[string][]*uint32),
 		animationFPS:      make(map[string]float64),
+		animationPlaying:  "",
 		animationEnabled:  false,
 	}
 }
@@ -40,14 +41,24 @@ func (bm *BasicMaterial) Render(delta float64, darkness float32) {
 	bm.UpdateAnimation(delta)
 	bm.UpdateAttribArrays()
 
-	gl.Uniform1f(bm.shader.GetUniform("diffuseLevel"), bm.diffuseLevel)
+	if bm.DiffuseMap != nil {
+		gl.ActiveTexture(gl.TEXTURE0)
+		gl.BindTexture(gl.TEXTURE_2D, *bm.DiffuseMap)
+	}
 
-	gl.Uniform3fv(bm.shader.GetUniform("color"), 1, &bm.color[0])
+	if bm.AlphaMap != nil {
+		gl.ActiveTexture(gl.TEXTURE1)
+		gl.BindTexture(gl.TEXTURE_2D, *bm.AlphaMap)
+	}
+
+	gl.Uniform1f(bm.shader.GetUniform("diffuseLevel"), bm.DiffuseLevel)
+
+	gl.Uniform4fv(bm.shader.GetUniform("hue"), 1, &bm.Hue[0])
 
 	gl.Uniform1i(bm.shader.GetUniform("diffuseMap"), 0)
-	gl.Uniform1f(bm.shader.GetUniform("diffuseMapScale"), bm.diffuseMapScale)
+	gl.Uniform1f(bm.shader.GetUniform("diffuseMapScale"), bm.DiffuseMapScale)
 
-	gl.Uniform1f(bm.shader.GetUniform("alphaLevel"), bm.alphaLevel)
+	gl.Uniform1f(bm.shader.GetUniform("alphaMapLevel"), bm.AlphaMapLevel)
 	gl.Uniform1i(bm.shader.GetUniform("alphaMap"), 1)
 
 	gl.Uniform1f(bm.shader.GetUniform("darkness"), darkness)
@@ -59,6 +70,10 @@ func (bm *BasicMaterial) UpdateAttribArrays() {
 	gl.EnableVertexAttribArray(2)
 }
 
+func (bm *BasicMaterial) GetShader() *ShaderProgram {
+	return bm.shader
+}
+
 func (bm *BasicMaterial) UpdateAnimation(delta float64) {
 	if bm.animationEnabled && bm.animationPlaying != "" {
 		if bm.animationFrame > 1/bm.animationFPS[bm.animationPlaying] {
@@ -68,9 +83,25 @@ func (bm *BasicMaterial) UpdateAnimation(delta float64) {
 				bm.animationCurrent = 0
 			}
 			bm.animationFrame = 0
-			bm.diffuseMap = bm.animationTextures[bm.animationPlaying][bm.animationCurrent]
+			bm.DiffuseMap = bm.animationTextures[bm.animationPlaying][bm.animationCurrent]
 		} else {
 			bm.animationFrame += delta
 		}
 	}
+}
+
+func (bm *BasicMaterial) EnableAnimation() {
+	bm.animationEnabled = true
+}
+
+func (bm *BasicMaterial) AddFrame(frame *uint32, anim string) {
+	bm.animationTextures[anim] = append(bm.animationTextures[anim], frame)
+}
+
+func (bm *BasicMaterial) SetAnimationFPS(anim string, s float64) {
+	bm.animationFPS[anim] = s
+}
+
+func (bm *BasicMaterial) PlayAnimation(anim string) {
+	bm.animationPlaying = anim
 }
