--- material/materialint.go
@@ -1,7 +1,0 @@
-package material
-
-type Material interface {
-	Render(delta float64, darkness float32)
-
-	GetShader() *ShaderProgram
-}
